fix path for log archives in nagios upgrade userguide
I just tried to follow this guide and ran into two problems. The first one is fairly obvious and fixed with this PR, s/archive/archives/.

The second one i'm not sure about. On my system all the nagios paths mentioned nagios3, not nagios. Might be a good idea to review and see if that should be fixed too.