make short and long service output internally consistent
I found it rather confusing that the numbers for "/" were different in
the short and long output examples.

It is not stated explicitly in the plugin guidelines if the information
in the short message should be duplicated into the long output, but I
guess it makes sense to do so.  My copy of check_disk does not even
support long service output :shrug: