make docker work
previously the dockerfile installed the dependencies and you then would mount
the source folder into the container, shadowing the just created node_modules
folder. So this setup only works if you had a node_modules folder on the host
machine already which was kind of the reason to use docker in first place.

To make this work, install the app into a subfolder and create the node_modules
on folder up.

Adopted some example files from here:

- https://github.com/BretFisher/node-docker-good-defaults

Also added a Makefile :-)

So you can just do a 'make docker-server' or 'make docker-build'.